vehicle/requirements.txt
kuksa-client
python-dotenv
